Location Validation

+++ server/api/location.controller.js
@@ -1,11 +1,16 @@
 require('rootpath')();
 const Q = require('q');
 const passport = require("passport");
+const validator = require('validator');
+const authLoginService = require('server/service/authlogin.service');
+const validatorService = require('server/service/validator.service');
 const locationservice = require ('server/service/location.service');
 const dataservice = require ('server/service/data.service');
 const knexconfig = require ('server/knexconfig.json');
 const knex = require('knex')(knexconfig);
 const express = require('express');
+const log4js = require('log4js');
+const logger = log4js.getLogger('location.controller');
 const router = express.Router();
 
 
@@ -16,14 +21,12 @@ router.get('/getlocations', passport.authenticate('jwt', { session: false }), ge
 router.get('/getlocationtypes', passport.authenticate('jwt', { session: false }), getlocationtypes);
 router.post('/addlocation', passport.authenticate('jwt', { session: false }), addlocation);
 router.delete('/:_id', passport.authenticate('jwt', { session: false }), deletelocation);
-router.put('/:_id', passport.authenticate('jwt', { session: false }), updatelocation);
-
-
-
+router.post('/:_id', passport.authenticate('jwt', { session: false }), updatelocation);
 
 
 
 module.exports = router;
+
 
 function getdojos(req,res) {
     
@@ -39,6 +42,7 @@ function getdojos(req,res) {
         res.status(400).send(err);
     });    
 }
+
 
 function getlocationnames(req,res) {
     
@@ -62,6 +66,7 @@ function getlocations(req,res) {
     }); 
 }
 
+
 function getlocationtypes(req,res) {
     locationservice.getlocationtypes()
     .then(function(locationtypes){
@@ -71,6 +76,8 @@ function getlocationtypes(req,res) {
         res.status(400).send(err);
     }); 
 }
+
+
 function addlocation(req, res){ //TO DO Validation
 
     if(true){ 
@@ -88,34 +95,77 @@ function addlocation(req, res){ //TO DO Validation
     }
 }
 
-function updatelocation(req, res){ //TO DO Validation
- 
 
-    if(true){ 
+function updatelocation(req, res){ 
+    logger.info(JSON.stringify(req.body));
 
-        locationservice.updatelocation(req)
-        .then(function(data){
-            console.log(JSON.stringify(data));
-          res.sendStatus(200).send(data);
+
+    var valid = !validator.isEmpty(JSON.stringify(req.body));
+    valid = validator.isJSON(JSON.stringify(req.body));
+    valid = valid && validator.isInt(req.body.ID.toString());
+    valid = valid && validator.isIn(req.body.locationtype, ['Dojo','Eseményhelyszín']);
+    valid = valid && validator.isDecimal(req.body.lat.toString());
+    valid = valid && validator.isDecimal(req.body.lon.toString());
+
+
+    var stringPromises = [  validatorService.validString(req.body.name, 50),
+                            validatorService.validString(req.body.city, 45),
+                            validatorService.validString(req.body.zipcode, 4),
+                            validatorService.validString(req.body.address, 45),
+                            validatorService.validString(req.body.building, 80)];
+    var allPromiseValid = false;
+
+    Q.allSettled(stringPromises)
+        .then(promiseResults => {
+            allPromiseValid =  promiseResults.reduce((all, elem) => all && elem.value, true);
+
         })
-        .catch(function (err) {
-            console.log(JSON.stringify(err));
-          res.sendStatus(400).send(err);
+        .then(() => authLoginService.getDecodedToken(req))
+        .then(token => validatorService.isDojocho(token.pid))
+        .then(isDojocho => {
+            if(valid && allPromiseValid && isDojocho) { 
+
+                locationservice.updatelocation(req)
+                .then(function(data){
+                    var message = ''.concat(data, ' helyszín frissítve.');
+                  res.sendStatus(204).send({message : message});
+                })
+                .catch(function (err) {
+                    logger.info('err: ' + JSON.stringify(err));
+                  res.sendStatus(400).send(err);
+                });
+            } else {
+                res.status(400).send({message:'Érvénytelen adat vagy nincs jogosultság!'});
+            }
+        })
+        .catch(err => {logger.info(JSON.stringify(err));
+            res.sendStatus(400).send(err);
         });
-    }else{
-        res.sendStatus(400).json({message:"Please send valid data."});
-        //Notify client to send valid data.
-    }
+
 }
 
 
-
 function deletelocation(req, res) {
-    locationservice.deletelocation(req.params._id)
-        .then(function () {
-            res.sendStatus(200);
+    
+    var valid = validator.isInt(req.params._id.toString());
+
+    authLoginService.getDecodedToken(req)
+        .then(token => {return validatorService.isDojocho(token.pid)})
+        .then(isDojocho => {
+
+            if(valid && isDojocho) {
+                locationservice.deletelocation(req.params._id)
+                .then(function () {
+                    res.sendStatus(200);
+                })
+                .catch(function (err) {
+                    res.status(400).send(JSON.stringify(err));
+                });
+            } else {
+                res.status(400).send({message : "Érvénytelen beérkező adat!"});
+            }
         })
-        .catch(function (err) {
-            res.status(400).send(err);
-        });
+        .catch(err => {
+            res.status(400).send(JSON.stringify(err));
+          });
 }
